requirements.txt
asgiref==3.8.1
blinker==1.4
command-not-found==0.3
cryptography==3.4.8
dbus-python==1.2.18
distlib==0.3.9
distro==1.7.0
distro-info==1.1+ubuntu0.1
Django==5.1.3
django-admin==2.0.2
django-excel-response2==3.0.6
django-six==1.0.5
djangorestframework==3.15.2
excel-base==1.0.4
filelock==3.16.1
greenlet==1.1.2
httplib2==0.20.2
importlib-metadata==4.6.4
isoweek==1.3.3
jeepney==0.7.1
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
more-itertools==8.10.0
msgpack==1.0.3
netifaces==0.11.0
numpy==1.26.4
oauthlib==3.2.0
pandas==2.2.2
platformdirs==4.3.6
PyGObject==3.42.1
PyJWT==2.3.0
pynvim==0.4.2
pyparsing==2.4.7
python-apt==2.4.0+ubuntu2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==5.4.1
screen==1.0.1
SecretStorage==3.3.1
six==1.16.0
sqlparse==0.5.1
systemd-python==234
TimeConvert==3.0.13
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
ubuntu-advantage-tools==8001
ufw==0.36.1
unattended-upgrades==0.1
virtualenv==20.27.1
wadllib==1.3.6
xlwt==1.3.0
zipp==1.0.0
